Fixed click events not propagating on perspective-viewer-hypergrid canvas if rendered initially with no data.
