+++ packages/perspective-viewer-hypergrid/test/js/regressions.spec.js
@@ -9,21 +9,7 @@
 
 const utils = require("@finos/perspective-viewer/test/js/utils.js");
 const path = require("path");
-
-async function capture_update(page, viewer, body) {
-    await page.evaluate(element => {
-        element.addEventListener("perspective-view-update", () => {
-            element.setAttribute("test-updated", true);
-        });
-    }, viewer);
-    await body();
-    try {
-        await page.waitFor(element => element.hasAttribute("test-updated"), {timeout: 3000}, viewer);
-    } catch (e) {
-        console.error("Missing 'test-updated' attribute");
-    }
-    await page.evaluate(element => element.removeAttribute("test-updated"), viewer);
-}
+const {capture_update} = require("./utils.js");
 
 utils.with_server({}, () => {
     describe.page(
